update free storage
