--- packages/free/src/shim.node.ts
@@ -1,1 +1,0 @@
-export { default as fetch } from 'node-fetch';
