+++ tsconfig.base.json
@@ -6,7 +6,7 @@
       "ES2019"
     ],
     "declaration": true,
-    "strict": false,
+    "strict": true,
     "esModuleInterop": true,
     "inlineSourceMap": true,
     "experimentalDecorators": true,
